Update 4.4 routing example to be more concise
It was somewhat confusing at first to switch examples midstream, where the initial basic controller example turns into a Form example when it comes time to passing an instantiated converted parameter. This commit aims to address this by providing a complete example of passing the converted node object into the first simple controller.